Fixed Fraction error, moved findFactors to its own utility class.
