+++ main/test/equation_builders/EquationDirectorTest.java
@@ -2,6 +2,7 @@ package equation_builders;
 
 import org.junit.Before;
 import org.junit.Test;
+import utilities.FactorFinder;
 
 import java.util.Arrays;
 
@@ -17,13 +18,13 @@ public class EquationDirectorTest {
 
     @Test
     public void testFindFactors() {
-        assertEquals("[1, 3, 35, 21, 5, 7, 105, 15]", Arrays.toString(EquationDirector.findFactors(105).toArray()));
+        assertEquals("[1, 3, 35, 21, 5, 7, 105, 15]", Arrays.toString(FactorFinder.findFactors(105).toArray()));
         //Second time should be identical
-        assertEquals("[1, 3, 35, 21, 5, 7, 105, 15]", Arrays.toString(EquationDirector.findFactors(105).toArray()));
+        assertEquals("[1, 3, 35, 21, 5, 7, 105, 15]", Arrays.toString(FactorFinder.findFactors(105).toArray()));
     }
 
     @Test
     public void testFindMultipleFactor() {
-        assertEquals("[128, 64, 32, 16, 1, 2, 4, 8]", Arrays.toString(EquationDirector.findFactors(128).toArray()));
+        assertEquals("[128, 64, 32, 16, 1, 2, 4, 8]", Arrays.toString(FactorFinder.findFactors(128).toArray()));
     }
 }
